pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apple-reminders-mcp"
version = "0.1.0"
description = "MCP server for managing Apple Reminders on macOS"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Shreyansh Jain", email = "shreyanshjain05@gmail.com" }
]
keywords = ["mcp", "apple", "reminders", "macos", "applescript", "ai", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastmcp>=2.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
apple-reminders-mcp = "apple_reminders_mcp.server:main"

[project.urls]
Homepage = "https://github.com/shreyanshjain05/apple_reminder_mcp_server"
Repository = "https://github.com/shreyanshjain05/apple_reminder_mcp_server"
Issues = "https://github.com/shreyanshjain05/apple_reminder_mcp_server/issues"

[tool.hatch.build.targets.wheel]
packages = ["apple_reminders_mcp"]
